pyproject.toml
[project]
name = "flask-demo"
version = "1.0.0"
dependencies = [
    "flask",
    "inject"
]

[project.optional-dependencies]
build = [
    "black",
    "isort",
    "mypy",
    "pylint"
]
test = [
    "pytest",
    "pytest-flask",
    "pytest-playwright"
]

[tool.isort]
profile = "black"

[tool.pylint.'MESSAGES CONTROL']
disable = [
    "missing-class-docstring",
    "missing-function-docstring",
    "missing-module-docstring"
]
good-names = ["id"]
